+++ conf/nginx.conf
@@ -48,18 +48,17 @@ http {
     add_header X-XSS-Protection          "1; mode=block";
     add_header X-Content-Type-Options    nosniff;
     add_header Content-Security-Policy   "default-src 'self' *.google.com *.amazon.com dap.digitalgov.gov *.google-analytics.com fonts.googleapis.com fonts.gstatic.com;";
-    index      index.html;
 
     server {
         client_body_temp_path /tmp/nginx_client_temp 1 2;
         listen                8080;
         listen                [::]:8080 default_server;
         root                  /usr/share/nginx/html;
-				client_max_body_size  5000000;
+        client_max_body_size  5000000;
 
         location / {
             autoindex on;
-            try_files $uri /index.html;
+            try_files $uri /public.html;
         }
 
         # Transparent proxy to Go API
@@ -70,6 +69,11 @@ http {
             proxy_set_header         Host $http_host;
             proxy_buffers            256 4k;
             proxy_max_temp_file_size 0k;
+        }
+
+        location ~ ^/(form|loading|token)/?(.*)$ {
+            autoindex on;
+            try_files $uri /form.html;
         }
     }
 }
