+++ src/form.html
@@ -1,0 +1,68 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="cache-control" content="no-cache, no-store, must-revalidate">
+    <meta name="pragma" content="no-cache">
+    <meta name="expires" content="0">
+    <title>eApp Prototype</title>
+
+    <!-- Icons -->
+    <link href="/img/favicons/favicon.ico"     rel="shortcut icon"                type="image/ico">
+    <link href="/img/favicons/favicon.png"     rel="icon"                         type="image/png">
+    <link href="/img/favicons/favicon-192.png" rel="icon"                         type="image/png" sizes="192x192">
+    <link href="/img/favicons/favicon-57.png"  rel="apple-touch-icon-precomposed" type="image/png">
+    <link href="/img/favicons/favicon-72.png"  rel="apple-touch-icon-precomposed" type="image/png" sizes="72x72">
+    <link href="/img/favicons/favicon-114.png" rel="apple-touch-icon-precomposed" type="image/png" sizes="114x114">
+    <link href="/img/favicons/favicon-144.png" rel="apple-touch-icon-precomposed" type="image/png" sizes="144x144">
+
+    <!-- HTML5 shim and Respond.js for IE8 support of HTML5 elements and media queries -->
+    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
+    <!--[if lt IE 9]>
+      <script type="text/javascript" src="/html5shiv.min.js"></script>
+      <script type="text/javascript" src="/respond.min.js"></script>
+    <![endif]-->
+    <link rel="stylesheet" type="text/css" href="/css/eqip.css" />
+  </head>
+  <body>
+    <table class="print-table">
+      <thead class="print-only">
+        <tr>
+          <th>
+            &nbsp;
+          </th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          <td>
+            <div id="app"></div>
+          </td>
+        </tr>
+      </tbody>
+      <tfoot class="print-only">
+        <tr>
+          <th>
+            &nbsp;
+          </th>
+        </tr>
+      </tfoot>
+    </table>
+    <div class="print-header print-only">FOR OFFICIAL USE ONLY</div>
+    <div class="print-footer print-only">FOR OFFICIAL USE ONLY</div>
+
+    <noscript>
+      <div class="usa-alert usa-alert-warning">
+        <div class="usa-alert-body">
+          <h3 class="usa-alert-heading">Your browser does not support JavaScript.</h3>
+          <p class="usa-alert-text">In order to view the full content of this site, please use a browser with JavaScript enabled. If you cannot, please <a href="https://github.com/18F/e-QIP-prototype/issues/new">file an issue on GitHub</a>.</p>
+        </div>
+      </div>
+    </noscript>
+
+    <!-- Include scripts at the bottom for performance gains in some browsers -->
+    <script type="text/javascript" language="javascript" src="/form.js"></script>
+  </body>
+</html>
