+++ src/public.html
@@ -63,7 +63,7 @@
     </noscript>
 
     <!-- Include scripts at the bottom for performance gains in some browsers -->
-    <script type="text/javascript" language="javascript" src="/eqip.js"></script>
+    <script type="text/javascript" language="javascript" src="/public.js"></script>
 
     <!-- We participate in the US government's analytics program. See the data at analytics.usa.gov. -->
     <script type="text/javascript" language="javascript" src="https://dap.digitalgov.gov/Universal-Federated-Analytics-Min.js" id="_fed_an_ua_tag"></script>
